separate page loader vs generic loader
